Updated dependencies - now installed via Composer
The used dependencies (Propel, a PHP ORM) were previously installed using git
submodules. However this doesn't install all necessary dependencies, so it is
changed to use Composer instead.

To install the dependencies Composer is used, which currently only loads the
Propel ORM with it's dependencies. To load the dependencies the command
  php composer.phar install
should be used. If composer is not installed, it can be obtained via
  curl -s http://getcomposer.org/installer | php